progress on using indexes.
in kv, both select/read calls now take a plan option.

most cases still pass None for the plan.

code to choose a plan based on the query is not there yet.

refactored listIndexes.

in most places where we need to choose an index, the list of
indexes does not exist and will need to be fetched.

indexes don't do dir<0 yet

fixed createIndex to automatically create collection if it
doesn't exist.

$min/$max index thing on find does use a plan now.  its parser
was also fixed to support multiple keys.