progress on indexing.
findPath now includes a BUndefined when constructing an array result for
a path that goes into an array of documents and one of them doesn't have
the key.  this is for indexing purposes.

questions about whether findPath should just return BUndefined instead
of an Option.  BUndefined is basically what None.

maybe BUndefined should never be written into bson, even though it can
be in the DU.

added removeUndefined so that these entries can be stripped out.

implemented chooseIndex.  very simplistic right now.  handles only
simple cases, and only EQ.

fixed the indexer to dive into arrays.  when an indexed value is
an array, it adds that entry, but it also adds one entry for every
item in the array.

still dealing with indexer issues involving null vs undefined.  might
need to add double entries in these cases.

an index of type text causes an error now instead of just being
ignored.  server3253 broken.

fixed createIndex to add entries for all documents that already exist
in the collection.