lots of changes to make lsm storage work.
list_indexes in the storage trait now accepts an optional db and coll
so it can constrain at the db level instead of requiring a filter
afterward.  this turns out to be a big perf win for both storage
layers.

fixed lsm flag 4 instead of 3.  oh my.

lsm RangeCursor no longer supports optional bounds.  that was silly.

fixed lsm problem where a root node landed on a boundary.  it
was creating a bad PageBlock.  current fix is that it requests another
block, which is it does not need, and then ends up giving it back.
should fix this.

fix lsm problem where writeOverflow was requesting a new segment
and that token was shadowing the one we were supposed to be using.

identified yet-to-be-fixed issue where lsm merge() will leave segments
forever in the 'merging' state when an error happens while trying to
write the merged segment.