simple and somewhat experimental range splitting for lsm.
define how many ranges.  keys are mapped to a range by linear
distribution on the first byte of the key.

added ConcatCursor, which is used to take all the SegmentCursors for a
given generation (one for each of the ranges), and view them as one
cursor.  MultiCursor now uses a stack of ConcatCursors instead of a
stack of SegmentCursors.

SegmentCursor doesn't open and own in its own file handle anymore.
all the subcursors for a given MultiCursor share a file handle.

changes to merge were minimal.  a merge takes place within a single
range.  each range is merged independently and has its own state.

all ranges within a level currently share one automerge thread.

need a layer to manage a pending transaction, now more than ever.
currently there is a hack in write_segment(), which is now mis-named
and should be called something like write_generation_of_segments(),
which allocates each key in the given BTreeMap to its proper
range.

overall perf impact of this change is positive, and would be even
more positive if the layer above (Elmo itself) had ways of being
more careful to confine its writes to one range.  need to explore
lsm allowing caller to specify key ranges in more detail.