Revert "fix merge level threshold computation to account for only the levels in the range being merged"
It turns out this is a problem.  Now each range can be merged
individually, we're gonna need different promotion thresholds
for each range.  Dividing the overall threshold by the number
of ranges does not work out well.  For now, applying the threshold to
the level as a whole turns out to be not the worst approach.

This reverts commit 700ef533e22ee0021d5a031237abe31286020a5f.