change merge from fresh to work in one of two ways.  single leaves are rewritten and bunched together.  anything with depth 1 or more is simply promoted to young without rewrite.
write_merge now allows behind to be optional.  so that merge
from fresh doesn't need to use it, because the only case where a
tombstone would be dropped during that kind of merge is when something
was committed and then immediately deleted before merge could move
it up.

when choosing something from one segment to promote to the next
level, allow the promotion of more than one page.  specifically,
the previous code would only allow promotion of a single leaf or
a single node of depth 1.  now we can promote multiple leaves (or
I suppose multiple nodes of depth higher) as long as the leaves
are contiguous children of the same parent node.  this required
MultiPageCursor to be used instead of PageCursor, which is fine,
as there should be no perf penalty for the case where MultiPageCursor
actually has just one page in it.

change merge from Young to be like merge from N to N+1.  instead
of always merging a whole segment regardless of size, choose a
set of (leaves) to promote.